Flask/requirements.txt
Flask==1.0.2
simplejson==3.16.0
spacy==2.0.18
tqdm==4.30.0
mysql-connector